작업 완료

+++ backend/src/main/java/com/example/backend/trma/dto/dataList/MoimDetailData.java
@@ -19,7 +19,7 @@ public class MoimDetailData {
     private Integer memberCnt;
     private String userId;
     private String userNm;
-    private int reviewScore;
+    private Double reviewScore;
     private String imageUrl;
     private String moimTitleEn;
     private String moimTitleJa;
